[PATCH 10/17] Fix / silence various compiler warnings
---
 interface/gtk_i.c    |   39 +++---
 interface/server_c.c |   12 +-
 interface/tk_c.c     |    7 +-
 interface/wrdt_tty.c |    4 +-
 interface/x_sherry.c |    4 +-
 interface/xaw_c.c    |   29 +++--
 interface/xaw_i.c    |   32 +++---
 interface/xskin_c.c  |   14 ++-
 libarc/unlzh.c       |    1 +
 libunimod/load_far.c |    4 +-
 libunimod/load_it.c  |    6 +-
 libunimod/mloader.c  |    4 +-
 libunimod/mlutil.c   |    4 +-
 timidity/ao_a.c      |    3 +-
 timidity/calcnewt.c  |   10 +-
 timidity/common.c    |   13 ++-
 timidity/effect.c    |    1 -
 timidity/esd_a.c     |    2 +-
 timidity/freq.c      |   14 +--
 timidity/instrum.c   |    3 +-
 timidity/loadtab.c   |    2 +-
 timidity/m2m.c       |   44 ++++----
 timidity/miditrace.c |    3 +-
 timidity/output.c    |    5 +-
 timidity/playmidi.c  |   19 ++--
 timidity/readmidi.c  |    5 +-
 timidity/recache.c   |    2 +-
 timidity/resample.c  |    2 +-
 timidity/reverb.c    |  336 +++++++++++++++++++++++++++++---------------------
 timidity/smplfile.c  |   10 +-
 timidity/speex_a.c   |    7 +-
 timidity/timidity.c  |   42 ++++---
 utils/net.c          |    2 +-
 utils/nkflib.c       |   18 ++-
 34 files changed, 391 insertions(+), 312 deletions(-)

+++ timidity/recache.c
@@ -59,7 +59,7 @@
 #define MAX_EXPANDLEN (1024 * 32)
 #define CACHE_DATA_LEN (allocate_cache_size / sizeof(sample_t))
 
-#define sp_hash(sp, note) ((unsigned int) (sp) + (unsigned int) (note))
+#define sp_hash(sp, note) ((unsigned long) (sp) + (unsigned int) (note))
 #define CACHE_RESAMPLING_OK 0
 #define CACHE_RESAMPLING_NOTOK 1
 #define SORT_THRESHOLD 20
